Add archive directory and move web folder to it
If applied, this commit will add the new top-level directory of archive to accomadate a wider range of data types without cluttering the root. The archive directory can be used for any data that does not fit into the other categories yet can still be categorised in some way. This will include websites archived for non-reading purposes, as well as datasets, backups of non-filetree data, as well as potentially creative assets.